pyproject.toml
[tool.poetry]
name = "feeding-jobstreet-be"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.12"
python-dotenv = "1.0.1"
Flask = "3.0.3"
gunicorn = "22.0.0"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
psycopg2-binary = "^2.9.9"
xlwt = "^1.3.0"
flask-cors = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
